+++ evolutionary_algorithm/details.py
