+++ evolutionary_algorithm/individuals/individual_structure.py
