+++ evolutionary_algorithm/individuals/traits/gene_sequence.py
