+++ evolutionary_algorithm/operators/base.py
@@ -2,7 +2,7 @@ from typing import Generic, TypeVar
 
 from typing_extensions import Final
 
-from evolutionary_programming.individuals import IndividualStructure
+from evolutionary_algorithm.individuals import IndividualStructure
 
 
 class InvalidOperatorError(ValueError):
