+++ evolutionary_algorithm/selectors/ranking.py
@@ -1,7 +1,7 @@
 from typing import Iterable, Sequence, Tuple, TypeVar, Generic
 
-from evolutionary_programming.details import Comparable
-from evolutionary_programming.individuals import IndividualType
+from evolutionary_algorithm.details import Comparable
+from evolutionary_algorithm.individuals import IndividualType
 
 ScoreType = TypeVar("ScoreType", bound=Comparable)
